# specs/000-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Manager API
  version: 0.3.0
  description: Contrato inicial do MVP. Endpoints autenticados usam sessão em cookie.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Iniciar sessão
      security: []
      x-requirements: [FR-001]
      parameters:
        - $ref: '#/components/parameters/AntiforgeryToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '204': { description: Sessão iniciada }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Encerrar sessão
      x-requirements: [FR-001]
      parameters:
        - $ref: '#/components/parameters/AntiforgeryToken'
      responses:
        '204': { description: Sessão encerrada }
        '400': { $ref: '#/components/responses/Problem' }
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentSession
      summary: Obter usuário, tenant, papel e permissões da sessão
      x-requirements: [FR-002, FR-027]
      responses:
        '200':
          description: Contexto sanitizado derivado da sessão atual
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentSession' }
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-waai-session
  parameters:
    AntiforgeryToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema: { type: string, minLength: 16 }
...
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    CurrentSession:
      type: object
      required: [user, tenant, role, permissions]
      properties:
        user: { $ref: '#/components/schemas/SessionUser' }
        tenant:
          anyOf:
            - $ref: '#/components/schemas/SessionTenant'
            - type: 'null'
        role: { type: string, enum: [PlatformAdmin, TenantOwner, Operator] }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
    SessionUser:
      type: object
      required: [id, email]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        officialApiLineCount: { type: integer, minimum: 0, nullable: true }
        qrCodeLineCount: { type: integer, minimum: 0, nullable: true }
    SessionTenant:
      type: object
      required: [id, name, slug, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [Active, Suspended] }
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        traceId: { type: string }
security:
  - sessionCookie: []